Check the status
Given two integer variables a and b, and a boolean variable flag. The task is to check the status and return accordingly.
Return True for the following cases:

Either a or b is non-negative and flag is false.
Both a and b are negative and flag is true.
Otherwise, return false.